refactor: use direct imports instead of lib imports
