+++ src/containers/Tenant/Query/utils/useSavedQueries.tsx
@@ -1,6 +1,7 @@
-import {SAVED_QUERIES_KEY, useSetting, useTypedSelector} from '../../../../lib';
 import {selectSavedQueriesFilter} from '../../../../store/reducers/queryActions/queryActions';
 import type {SavedQuery} from '../../../../types/store/query';
+import {SAVED_QUERIES_KEY} from '../../../../utils/constants';
+import {useSetting, useTypedSelector} from '../../../../utils/hooks';
 
 export function useSavedQueries() {
     const [savedQueries] = useSetting<SavedQuery[]>(SAVED_QUERIES_KEY, []);
